moved more security related configurations to the separate module
