+++ modules/flowable-ui-idm/flowable-ui-idm-app/pom.xml
@@ -26,7 +26,7 @@
         </dependency>
         <dependency>
             <groupId>org.flowable</groupId>
-            <artifactId>flowable-ui-idm-logic-impl</artifactId>
+            <artifactId>flowable-ui-idm-security</artifactId>
         </dependency>
         <dependency>
             <groupId>org.flowable</groupId>
