+++ modules/flowable-ui-idm/flowable-ui-idm-rest/pom.xml
@@ -19,7 +19,7 @@
         </dependency>
         <dependency>
             <groupId>org.flowable</groupId>
-            <artifactId>flowable-ui-idm-logic-impl</artifactId>
+            <artifactId>flowable-ui-idm-security</artifactId>
         </dependency>
 		<!-- Flowable -->
 		<dependency>
