+++ modules/flowable-ui-idm/flowable-ui-idm-security/src/main/java/org/flowable/app/security/AjaxLogoutSuccessHandler.java
