+++ modules/flowable-ui-idm/pom.xml
@@ -28,7 +28,7 @@
 	
 	<modules>
 		<module>flowable-ui-idm-logic</module>
-                <module>flowable-ui-idm-logic-impl</module>
+                <module>flowable-ui-idm-security</module>
 		<module>flowable-ui-idm-rest</module>
 		<module>flowable-ui-idm-conf</module>
 		<module>flowable-ui-idm-app</module>
@@ -43,7 +43,12 @@
         </dependency>
         <dependency>
             <groupId>org.flowable</groupId>
-            <artifactId>flowable-ui-idm-logic-impl</artifactId>
+            <artifactId>flowable-ui-idm-security</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.seal.process</groupId>
+            <artifactId>seal-flowable-ui-idm-logic-impl</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
